{{cookiecutter.repo_name}}/pyproject.toml
[tool.poetry]
name = "{{cookiecutter.repo_name}}"
version = "{{cookiecutter.version}}"
description = "{{cookiecutter.short_description}}"
authors = ["{{cookiecutter.full_name}} <{{cookiecutter.email}}>"]
license = "MIT"

keywords = ["sitri", "console", "cli", "utils", "typer"]

repository = "https://github.com/{{cookiecutter.github_username}}/{{cookiecutter.repo_name}}"
documentation = "https://{{cookiecutter.repo_name}}.readthedocs.io"

classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Other Environment",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
]

[tool.poetry.dependencies]
python = "==3.10"

typer = "^0.4.1"
sitri = {version = "^2.0.0", extras = ["pyyaml"]}
PyYAML = "^6.0"
pydantic = "^1.9.0"
"github3.py" = "^3.2.0"
structlog = "^21.5.0"
cbor2 = "^5.4.2"
ujson = "^5.2.0"
types-PyYAML = "^6.0.7"
art = "^5.6"
yaspin = "^2.1.0"
tabulate = "^0.8.9"

[tool.poetry.dev-dependencies]
pre-commit = "^2.18.1"
commitizen = "^2.24.0"
pytype = "^2022.4.26"
mypy = "^0.950"

[tool.black]
line-length = 120
target-version = ['py310']
include = '\.pyi?$'
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | foo.py           # also separately exclude a file named foo.py in
                     # the root of the project
)
'''

[tool.isort]
known_first_party = "{{cookiecutter.repo_name}}"
line_length = 120
profile = "black"
multi_line_output = 3
include_trailing_comma = true

[tool.pycln]
all = true

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
no_site_packages = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
